added click login method with string parameter
